% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt.R
\name{register_dt}
\alias{register_dt}
\title{Register a DT DataTable Component}
\usage{
register_dt(
  registry,
  dt_output_id,
  data_reactive,
  shared_id_column,
  click_handler = NULL
)
}
\arguments{
\item{registry}{A link registry created by \code{create_link_registry()}}

\item{dt_output_id}{Character string: the outputId of your DT::DTOutput}

\item{data_reactive}{Reactive expression returning the data frame for the table}

\item{shared_id_column}{Character string: name of the ID column}
}
\description{
Register a DT datatable for linking with other components.
}
