% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly.R
\name{prepare_plotly_linking}
\alias{prepare_plotly_linking}
\title{Prepare Plotly for Linking}
\usage{
prepare_plotly_linking(plotly_obj, id_column, source)
}
\arguments{
\item{plotly_obj}{A plotly object created with plot_ly()}

\item{id_column}{Character string: name of the ID column in the data}

\item{source}{Character string: plotly source identifier}
}
\value{
Modified plotly object with linking parameters added
}
\description{
Utility function to automatically add required parameters to a plotly object
for reliable linking, regardless of plot structure (single/multiple traces).
}
\examples{
\dontrun{ TODO (update example to actually work)
# Instead of manually adding customdata/key:
p <- plot_ly(data = my_data, x = ~x_col, y = ~y_col, color = ~category)
p <- prepare_plotly_linking(p, "my_id_column", "my_source")
}
}
